Split app flow presentation
